requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.1
python-multipart>=0.0.9
pydantic>=2.6.3
pydantic-settings>=2.2.1
langchain>=0.1.12
langchain-core
langchain-community
langchain-text-splitters
langchain-chroma
sentence-transformers
chromadb>=0.4.24
pypdf>=4.1.0
python-dotenv>=1.0.1
google-genai
pytest>=8.0.2